Inform user if a particular datacast message is received.
The following patch send a message to user if a particular type of
datacast is received, for example if A send a wink (that isn't
supported by pecan) to B (that uses msn-pecan), B will receive the
following message instead of a nudge:

A sent you a wink, but it isn't supported yet, so it cannot be viewed.

Really useful to know if a WLM contact is sending you a voice clip or
a wink (actually, if it does you receive only a nudge).